REFACTOR/split Landing.tsx to multiple components improving readability and maintainability
