+++ src/styles/Landing.css
@@ -1,0 +1,129 @@
+// small inline CSS to keep everything inside the file (glass, float, snap fine-tuning)
+
+/* float */
+@keyframes floatY { 
+  0% { 
+    transform: translateY(0); 
+  } 
+  50% { 
+    transform: translateY(-8px);
+  } 
+  100% { 
+    transform: translateY(0);
+  } 
+}
+
+.float-slow { 
+  animation: floatY 6s ease-in-out infinite; 
+}
+
+/* fade-in helper */
+.fade-in { 
+  opacity: 0; 
+  transform: translateY(8px); 
+  transition: opacity .45s ease, transform .45s cubic-bezier(.2,.9,.2,1); 
+  will-change: opacity, transform; 
+}
+
+.fade-in.in-view { 
+  opacity: 1; 
+  transform: translateY(0); 
+}
+
+/* small nav outline removal for the centered nav look */
+.nav-outline { 
+  box-shadow: 0 8px 32px rgba(2,6,23,0.45); 
+  border: 1px solid rgba(255,255,255,0.04); 
+}
+
+/* make anchor jumps smooth and avoid tiny flicker */
+html { 
+  scroll-behavior: smooth; 
+}
+
+/* Subtle Shine Effect for professional polish */
+.shine-effect { 
+  position: relative; 
+  overflow: hidden;
+  transition: all 0.3s ease;
+}
+
+.shine-effect::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: -100%;
+  width: 50%;
+  height: 100%;
+  background: linear-gradient(90deg, rgba(255, 255, 255, 0) 0%, rgba(255, 255, 255, 0.1) 50%, rgba(255, 255, 255, 0) 100%);
+  transform: skewX(-30deg);
+  transition: all 0.5s ease;
+}
+
+.shine-effect:hover::before {
+  left: 150%;
+}
+
+/* Mobile Menu Hover Effects - Desktop only */
+@media (hover: hover) and (pointer: fine) {
+  .menu-item:hover {
+    transform: scale(1.02);
+    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1)
+  }
+  .menu-item:hover .menu-item-icon {
+    transform: scale(1.1) rotate(3deg);
+    background: linear-gradient(135deg, rgba(20, 184, 166, 0.2), rgba(6, 182, 212, 0.2));
+  }
+  .menu-item:hover .menu-item-icon-inner {
+    color: rgb(94 234 212);
+  }
+  
+  .menu-item:hover .menu-item-text {
+    transform: translateX(8px);
+  }
+      
+  .menu-item:hover .menu-item-arrow {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+  
+.menu-item-arrow {
+  opacity: 0;
+  transform: translateX(-8px);
+}
+  
+/* Mobile touch feedback */
+.menu-item.touch-active {
+  transform: scale(0.98);
+  box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+}
+  
+.menu-item.touch-active .menu-item-icon {
+  transform: scale(0.95);
+  background: linear-gradient(135deg, rgba(20, 184, 166, 0.3), rgba(6, 182, 212, 0.3));
+}
+  
+.menu-item.touch-active .menu-item-icon-inner {
+  color: rgb(94 234 212) !important;
+}
+  
+.menu-item.touch-active .menu-item-text {
+  transform: translateX(4px);
+}
+  
+.menu-item.touch-active .menu-item-arrow {
+  opacity: 0.8;
+  transform: translateX(2px);
+}
+  
+/* For light mode adjustments */
+.light .menu-item:hover .menu-item-icon-inner,
+.light .menu-item.touch-active .menu-item-icon-inner {
+  color: rgb(13 148 136);
+}
+  
+/* Card glow on hover (dark mode only) */
+.card-glow-hover:hover {
+  box-shadow: 0 0 40px rgba(52, 211, 163, 0.15), 0 0 10px rgba(52, 211, 163, 0.1);
+}
